# Use the decremental approach in the hypervolume contribution calculation