--- a/optuna/samplers/_tpe/sampler.py
+++ b/optuna/samplers/_tpe/sampler.py
@@ -835,9 +835,18 @@ def _calculate_weights_below_for_multi_objective(
 
     loo_mat = ~np.eye(pareto_sols.shape[0], dtype=bool)  # Leave-one-out bool matrix.
     contribs = np.zeros(n_below_feasible, dtype=float)
-    contribs[on_front] = hv - np.array(
-        [compute_hypervolume(pareto_sols[loo], ref_point, assume_pareto=True) for loo in loo_mat]
-    )
+    if len(study.directions) <= 3:
+        contribs[on_front] = [
+            hv - compute_hypervolume(pareto_sols[loo], ref_point, assume_pareto=True)
+            for loo in loo_mat
+        ]
+    else:
+        hvs = []
+        for i in range(pareto_sols.shape[0]):
+            limited_sols = np.maximum(pareto_sols[loo_mat[i, :]], pareto_sols[i, :])
+            intersec_hv = compute_hypervolume(limited_sols, ref_point)
+            hvs.append(np.prod(ref_point - pareto_sols[i, :]) - intersec_hv)
+        contribs[on_front] = hvs
     weights_below[is_feasible] = np.maximum(contribs / max(np.max(contribs), EPS), EPS)
     return weights_below