--- a/optuna/samplers/_tpe/sampler.py
+++ b/optuna/samplers/_tpe/sampler.py
@@ -844,8 +844,7 @@ def _calculate_weights_below_for_multi_objective(
         contribs[on_front] = np.prod(ref_point - pareto_sols, axis=-1)
         limited_sols = np.maximum(pareto_sols, pareto_sols[:, np.newaxis])
         contribs[on_front] -= [
-            compute_hypervolume(limited_sols[i, loo], ref_point)
-            for i, loo in enumerate(loo_mat)
+            compute_hypervolume(limited_sols[i, loo], ref_point) for i, loo in enumerate(loo_mat)
         ]
     weights_below[is_feasible] = np.maximum(contribs / max(np.max(contribs), EPS), EPS)
     return weights_below