@nabenabe0928 2025-09-02
Related:
- https://github.com/optuna/optuna/pull/6232
- https://github.com/optuna/optuna/pull/6234